pyproject.toml
[tool.poetry]
name = "page-summarizer"
version = "0.1.0"
description = "A command-line tool to summarize web pages and YouTube videos."
authors = ["Anthony Scotti"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "2.31.0"
python-dotenv = "1.0.0"
langchain = "0.0.340"
openai = "1.3.5"
html2text = "2020.1.16"
youtube-transcript-api = "0.6.1"

[tool.poetry.dev-dependencies]
ruff = "0.1.9"

[tool.ruff]
select = ["E", "F", "Q", "I"]
ignore = ["E501"]
